src/discord_bot/requirements.txt
discord.py
openai
schedule
python-dotenv
pytz
